fixed #4

+++ src/main/java/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
     double calculate(double masskg, double growthm) {
-        int bmi = (int) (masskg / (growthm / 100 * growthm / 100));
+        double bmi = (double) (masskg / (growthm / 100 * growthm / 100));
         return bmi;
     }
 }
